crypto: move helper.h up a directory
